Moved events components to where it belongs
